requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pg8000==1.31.2
python-dotenv==1.0.1
pydantic==2.10.3
google-generativeai==0.8.3
